requirements.txt
dash>=4.0
plotly>=6.0
pandas>=2.0
numpy>=1.24
yfinance>=0.2
simfin>=0.6
finnhub-python>=2.4
defeatbeta-api>=0.0.50
pyarrow>=15.0
openpyxl>=3.1
requests>=2.31
